+++ R/plotting.R
@@ -25,39 +25,39 @@ get_ggplot <- function(x, color_group, color_id, facet_group, facet_id, facet_ty
 #       "pith/bark" into a legend.
   stopifnot(facet_type %in% c("grid", "wrap"))
   stopifnot(rugbuffer_size >= 2)
-  clean.nonrec <- subset(x$rings, x$rings$type != "recorder.year")
+  clean.nonrec <- subset(x$rings, x$rings$rec_type != "recorder.year")
   scar.types <- c("unknown.fs", "dormant.fs", "early.fs",
                   "middle.fs", "late.fs", "latewd.fs")
   injury.types <- c("unknown.fi", "dormant.fi", "early.fi",
                     "middle.fi", "late.fi", "latewd.fi")
   pithbark.types <- c("pith.year", "bark.year")
-  events <- subset(clean.nonrec, (type %in% scar.types) | (type %in% injury.types) | (type %in% pithbark.types))
-  levels(events$type)[levels(events$type) %in% scar.types] <- "Scar"
-  levels(events$type)[levels(events$type) %in% injury.types] <- "Injury"
-  levels(events$type)[levels(events$type) %in% pithbark.types] <- "Pith/Bark"
-  events$type <- factor(events$type, levels = c("Scar", "Injury", "Pith/Bark"))
+  events <- subset(clean.nonrec, (rec_type %in% scar.types) | (rec_type %in% injury.types) | (rec_type %in% pithbark.types))
+  levels(events$rec_type)[levels(events$rec_type) %in% scar.types] <- "Scar"
+  levels(events$rec_type)[levels(events$rec_type) %in% injury.types] <- "Injury"
+  levels(events$rec_type)[levels(events$rec_type) %in% pithbark.types] <- "Pith/Bark"
+  events$rec_type <- factor(events$rec_type, levels = c("Scar", "Injury", "Pith/Bark"))
   
   live <- aggregate(x$rings$year, by = list(x$rings$series), FUN = range, na.rm = TRUE)
   live <- data.frame(series = live$Group.1,
                      first = live$x[, 1],
                      last = live$x[, 2],
-                     type = rep("non-recording", dim(live)[1]))
-  recorder <- subset(x$rings, x$rings$type == "recorder.year")
+                     rec_type = rep("non-recording", dim(live)[1]))
+  recorder <- subset(x$rings, x$rings$rec_type == "recorder.year")
   if ( dim(recorder)[1] > 0 ) {  # If there are recorder years...
     # Get the min and max of the recorder years.
     recorder <- aggregate(recorder$year,  # TODO: rename this var.
-                           by = list(recorder$series, recorder$type),
+                           by = list(recorder$series, recorder$rec_type),
                            FUN = range,
                            na.rm = TRUE)
     recorder <- data.frame(series = recorder$Group.1,
                            first = recorder$x[, 1],
                            last = recorder$x[, 2],
-                           type = rep("recording", dim(recorder)[1]))
+                           rec_type = rep("recording", dim(recorder)[1]))
     segs <- rbind(recorder, live)
   } else {  # If there are no recorder years...
     segs <- live
   }
-  levels(segs$type) <- c("Recording", "Non-recording")
+  levels(segs$rec_type) <- c("Recording", "Non-recording")
   
   p <- NA
   rings <- x$rings
@@ -76,11 +76,11 @@ get_ggplot <- function(x, color_group, color_id, facet_group, facet_id, facet_ty
                     by = "series")
     p <- ggplot2::ggplot(rings, ggplot2::aes(y = series, x = year, color = species))
   }
-  p <- (p + ggplot2::geom_segment(ggplot2::aes(x = first, xend = last, y = series, yend = series, linetype = type),
+  p <- (p + ggplot2::geom_segment(ggplot2::aes(x = first, xend = last, y = series, yend = series, linetype = rec_type),
                          data = segs)
           + ggplot2::scale_linetype_manual(values = c("solid", "dashed", "solid")))
           #+ ggplot2::scale_size_manual(values = c(0.5, 0.5, 0.3)))
-  p <- (p + ggplot2::geom_point(data = events, ggplot2::aes(shape = type, size = type),
+  p <- (p + ggplot2::geom_point(data = events, ggplot2::aes(shape = rec_type, size = rec_type),
                        #size = event_size, color = "black")
                        color = "black")
           + ggplot2::scale_size_manual(values = event_size)